Create GC_content.kt
GC-content is an important feature of the genome sequences and is defined as the percentage ratio of the sum of all guanines and cytosines to the overall number of nucleic bases in the genome sequence.

Write a program, which calculates the percentage of G characters (guanine) and C characters (cytosine) in the entered string. Your program should be case independent.
Note, answer should be double.

Remember: characters are case-insensitive(g equals G and c equals C)

Input data format

The single input line contains a sequence.